move loading/writing of notes
